AI_interviewer/backstage/requirements_agent.txt
# AI Interviewer Backend - Agent 模块依赖
# 在原有 requirements_clean.txt 基础上添加以下依赖

# ==================== LangGraph 相关 ====================
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

# ==================== 向量嵌入相关 ====================
numpy>=1.24.0
dashscope>=1.20.0

# ==================== Web 搜索相关 ====================
# 以下搜索 API 选择一个配置即可
# Tavily (推荐，需要 API Key)
tavily-python>=0.5.0

# DuckDuckGo (免费后备方案)
duckduckgo-search>=6.0.0

# aiohttp (用于 Serper API)
aiohttp>=3.9.0

# ==================== 原有依赖 ====================
aiomysql>=0.3.2
anyio>=4.0.0
async-timeout>=5.0.1
cffi>=2.0.0
click>=8.1.8
colorama>=0.4.6
cryptography>=46.0.0
fastapi>=0.109.0
greenlet>=3.2.0
h11>=0.16.0
idna>=3.4
pycparser>=2.23
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyMySQL>=1.1.2
python-dotenv>=1.0.0
redis>=7.0.0
SQLAlchemy>=2.0.0
starlette>=0.32.0
typing_extensions>=4.0.0
uvicorn>=0.28.0
python-jose[cryptography]>=3.3.0
openai>=1.0.0
